Added three emojis: England, Scotland, Wales
England, Scotland, and Wales display as single character emojis, and are the only single flag country emojis. As such, they fulfill the qualifications for a base single emoji, as they are not unlike the pirate flag (🏴‍☠️). The UK flags are classified as RGI under the Unicode spec, and are fully supported as RGI emoji across multiple vendors.

More info:
https://unicode.org/emoji/charts/full-emoji-list.html#subdivision-flag
https://adraffy.github.io/ens-normalize.js/test/resolver.html